refactor: move data to section
